app/requirements.txt
PyQt6>=6.0.0
Pillow>=9.0.0
imagehash>=4.3.1
numpy>=1.21.0
pillow-avif-plugin>=1.0.0